+++ lib/Table/TableAutomaticStylesStratum.ts
@@ -1,6 +1,7 @@
 import i18next from "i18next";
 import { uniq } from "lodash-es";
 import { computed } from "mobx";
+import { createTransformer } from "mobx-utils";
 import isDefined from "../Core/isDefined";
 import { JsonObject } from "../Core/Json";
 import ConstantColorMap from "../Map/ColorMap/ConstantColorMap";
@@ -96,7 +97,7 @@ export default class TableAutomaticStylesStratum extends LoadableStratum(
           idColumns: idColumn && [idColumn.name]
         }),
         color: createStratumInstance(TableColorStyleTraits, {
-          legend: new ColorStyleLegend(this.catalogItem, -1)
+          legend: this._createLegendForColorStyle(-1)
         })
       });
     }
@@ -110,7 +111,7 @@ export default class TableAutomaticStylesStratum extends LoadableStratum(
     // Just add default legend
     return createStratumInstance(TableStyleTraits, {
       color: createStratumInstance(TableColorStyleTraits, {
-        legend: new ColorStyleLegend(this.catalogItem, -1)
+        legend: this._createLegendForColorStyle(-1)
       })
     });
   }
@@ -130,7 +131,7 @@ export default class TableAutomaticStylesStratum extends LoadableStratum(
     if (scalarColumns.length >= (hasTime ? 1 : 2)) {
       return createStratumInstance(TableStyleTraits, {
         color: createStratumInstance(TableColorStyleTraits, {
-          legend: new ColorStyleLegend(this.catalogItem, -1)
+          legend: this._createLegendForColorStyle(-1)
         }),
         chart: createStratumInstance(TableChartStyleTraits, {
           xAxisColumn: hasTime ? timeColumns[0].name : scalarColumns[0].name,
@@ -159,7 +160,7 @@ export default class TableAutomaticStylesStratum extends LoadableStratum(
         id: column.name,
         color: createStratumInstance(TableColorStyleTraits, {
           colorColumn: column.name,
-          legend: new ColorStyleLegend(this.catalogItem, i)
+          legend: this._createLegendForColorStyle(i)
         }),
         pointSize: createStratumInstance(TablePointSizeStyleTraits, {
           pointSizeColumn: column.name
@@ -230,6 +231,12 @@ export default class TableAutomaticStylesStratum extends LoadableStratum(
       return `${this.catalogItem.activeTableStyle.timeColumn.title}: `;
     }
   }
+
+  private readonly _createLegendForColorStyle = createTransformer(
+    (i: number) => {
+      return new ColorStyleLegend(this.catalogItem, i);
+    }
+  );
 }
 
 export class ColorStyleLegend extends LoadableStratum(LegendTraits) {